refactor, fix docs
