+++ specification/inference/put_openai/PutOpenAiRequest.ts
@@ -17,13 +17,13 @@
  * under the License.
  */
 
+import { InferenceChunkingSettings } from '@inference/_types/Services'
 import {
-  InferenceChunkingSettings,
-  RateLimitSetting
-} from '@inference/_types/Services'
+  OpenAIServiceSettings,
+  OpenAITaskSettings
+} from '@inference/_types/Settings'
 import { RequestBase } from '@_types/Base'
 import { Id } from '@_types/common'
-import { integer } from '@_types/Numeric'
 
 /**
  * Create an OpenAI inference endpoint.
@@ -61,7 +61,7 @@ export interface Request extends RequestBase {
   }
   body: {
     /**
-     * The chunking configuration object.
+     * The chunking configuration object. Only applies for the `text_embedding` task.
      * @ext_doc_id inference-chunking
      */
     chunking_settings?: InferenceChunkingSettings
@@ -89,56 +89,4 @@ export enum OpenAITaskType {
 
 export enum ServiceType {
   openai
-}
-
-export class OpenAIServiceSettings {
-  /**
-   * A valid API key of your OpenAI account.
-   * You can find your OpenAI API keys in your OpenAI account under the API keys section.
-   *
-   * IMPORTANT: You need to provide the API key only once, during the inference model creation.
-   * The get inference endpoint API does not retrieve your API key.
-   * After creating the inference model, you cannot change the associated API key.
-   * If you want to use a different API key, delete the inference model and recreate it with the same name and the updated API key.
-   * @ext_doc_id openai-api-keys
-   */
-  api_key: string
-  /**
-   * The number of dimensions the resulting output embeddings should have.
-   * It is supported only in `text-embedding-3` and later models.
-   * If it is not set, the OpenAI defined default for the model is used.
-   */
-  dimensions?: integer
-  /**
-   * The name of the model to use for the inference task.
-   * Refer to the OpenAI documentation for the list of available text embedding models.
-   * @ext_doc_id openai-models
-   */
-  model_id: string
-  /**
-   * The unique identifier for your organization.
-   * You can find the Organization ID in your OpenAI account under *Settings > Organizations*.
-   */
-  organization_id?: string
-  /**
-   * This setting helps to minimize the number of rate limit errors returned from OpenAI.
-   * The `openai` service sets a default number of requests allowed per minute depending on the task type.
-   * For `text_embedding`, it is set to `3000`.
-   * For `completion`, it is set to `500`.
-   */
-  rate_limit?: RateLimitSetting
-  /**
-   * The URL endpoint to use for the requests.
-   * It can be changed for testing purposes.
-   * @server_default https://api.openai.com/v1/embeddings.
-   */
-  url?: string
-}
-
-export class OpenAITaskSettings {
-  /**
-   * For a `completion` or `text_embedding` task, specify the user issuing the request.
-   * This informaiton can be used for abuse detection.
-   */
-  user?: string
 }
